// src/types/videoTypes.ts
export type VideoDetails = {
  id: string;
  Title: string;
  description: string;
  course_id: string;
  parent_id: number;
  duration: string;
  date_and_time: string;
  event_date: string;
  exam: string;
  section: number;
  subject: number;
  topic: number;
  concept: number;
  filter: string;
  filter_name: string;
  folder_wise_course: number;
  app_status: string;
  chat_status: string;
  free_flag: number;
  is_purchased: string;
  is_premiere: string;
  is_pdf_encrypted: string;
  is_pdf2_encrypted: string;
  open_outside_flag: string;
  enable_video_download: string;
  enable_video_restrictions: number;
  low_latency_enabled: boolean;
  live_status: number;
  live_type: string;
  live_rewind_enable: string;
  save_flag: string;
  videos_count: string;
  images_count: string;
  files_count: string;
  quiz_count: string;
  tests_count: string;
  quiz_title_id: string;
  test_title_id: string;
  live_quiz_id: number;
  key: string;
  iv_string: string;
  video_key: string;
  video_id: string;
  cookie_key: string;
  cookie_value: string;
  pdf_encryption_key: string;
  pdf_encryption_version: string;
  pdf2_encryption_key: string;
  pdf2_encryption_version: string;
  path: string;
  quality: string;
  file_link: string;
  pdf_link: string;
  pdf_link2: string;
  study_material_link: string;
  download_link: string;
  download_link2: string;
  download_url_higher_version: string;
  download_url_lower_version: string;
  video_player_url: string;
  video_player_lower_url: string;
  video_player_token: string;
  embed_url: string;
  thumbnail: string;
  recording_hls: string;
  recording_schedule: string;
  recording_type: string;
  rec_domain: string;
  hls_stream_type: string;
  uhs_version: string;
  ytFlag: number;
  ytFlagWeb: number;
  strtotime: number;
  sortingparam: string;
  encrypted_links: {
    quality: string;
    path: string;
    backup_url: string;
    backup_url2: string;
    key: string;
  }[];
  livestream_links: any[];
  download_links: any[];
  webdrm_links: any[];
  links: any[];
  meeting_id: string;
  meeting_password: string;
  special_course: string | null;
  max_time_allowed: string;
};

export type VideoDetailsResponse = {
  message: string;
  status: number;
  data: VideoDetails;
};